Minor changes
Just some minor changes so the gui works with the new script release. I
changed some of the option names to better align with the rest of the
script. Also removed the GUI-Script communication code. Finally, a bug
existed that didn't reset the change data on loading a new file which
caused new files with no changes to be detected as files with changes.
The script would therefore always ask to save changes when loading a new
file, even if the old one had no changes. This is fixed now.