Updated GUI with missing items
Forgot to add max_dl_tries and retry_wait to the GUI, so that had to be
done. I was pleasantly surprised by how easy it was to add new items.
The most complicated part was adding in the new defaults, something I
think I could improve in the future by not using the one-liner approach
(1800+ char lines kill kitties). I also fixed a handfull of bugs that I
hadn't noticed before. There was an issue if you added a bunch of
watch/avoids without actually clicking on them to edit them (who does
that? me lol). There was also a problem with case-sensitivity in
handling variables in the scc2.ini. Options in scc2.ini should not be
case sensitive anymore. Preventing accidentally setting checkboxes as
tristate if the user entered a 1 instead of a 2 in the ini file. Now all
1's are autocorrected to 2's for non-stristate checkboxes. On occasion
the GUI could ask you twice if you wanted to save or discard changes,
usually when trying to load a file after creating a new one from scratch
but not saving. This has been fixed, which also cleaned up the code a
little (see clearGUIState() function). Finally all watchlist/avoidlist
items now have their title set by default at creation, instead of when
you click in the name textbox.