Fixed cut off uT mode description
Never noticed before but all the info was cut off. Moved the label to
the left and slightly down while widening it enough to fully display the
uT mode. Updated the version checking function since it was not working
with double digit minor build numbers. New function looks ugly but it
works so meh. Also made FTP PASV mode default to on. Nearing final
release.  :)