Basically done!
Added in file drag-and-drop support (I was getting tired of using the
menu to load configs while testing :p) and updated the email/command
shortcut list, changing %tag% to %watchname%. If nobody finds any bugs
in this one It'll officially be done! YAY! I tried to reproduce the bug
that someone reported where loading configs in windows made in linux (or
vice-versa) would cause errors but I was unsuccessful. If this issue
comes back again I know how to fix it but so far its unnecessary.