Undo/Redo v0.1
Proof of concept for the Undo/Redo system. All QLineEdits in general
settings have the undo/redo system working in them. Had to re-implement
a few things to get it working but I think this system is going to work
out. Didn't want to commit this until I was finished but this is a good
place to do it anyway.