Starting work on v2.2
Getting started on version 2.2 of the script and GUI. This version will
allow administering of the script from a remote box. This version will
also include the ability to add tags to a release loaded in uTorrent as
well as a utorrent rss downloader style episode filter to prevent
multiple copies of the same episode in different qualities from
downloading. Current idea is to download only a higher quality release.
So if the first one to be autodownloaded is 720p, no lower quality
versions will be downloaded however a 1080p release will be.

This specific release has a ton of bugs, just a heads up.